lab_02/app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
redis==5.0.1
pydantic==2.4.2
python-dotenv==1.0.0

